2. Тесты для идеальный вес [#342009]

+++ src/ru/job4j/calculator/Fit.java
@@ -3,14 +3,14 @@ package ru.job4j.calculator;
 public class Fit {
 
     // Идеальный вес для мужчин = (рост в сантиметрах – 100) · 1,15.
-    public static double manWeight(short height) {
+    public static double manWeight(double height) {
         double rsl = (height - 100) * 1.15;
         return rsl;
     }
 
     // Идеальный вес для женщин = (рост в сантиметрах – 110) · 1,15.
     @SuppressWarnings("checkstyle:WhitespaceAround")
-    public static double womanWeight(short height) {
+    public static double womanWeight(double height) {
         double rsl = (height - 110) * 1.15;
         return rsl;
     }
